Move data manipulation functions out of components.
