+++ static/js/publisher/release/historyPanel.js
@@ -9,6 +9,7 @@ export default class HistoryPanel extends Component {
       <div className="p-history-panel">
         <div className="p-strip is-shallow">
           <RevisionsList
+            releases={this.props.releases}
             revisionsMap={this.props.revisionsMap}
             revisionsFilters={this.props.revisionsFilters}
             releasedChannels={this.props.releasedChannels}
@@ -17,7 +18,6 @@ export default class HistoryPanel extends Component {
             showChannels={true}
             showArchitectures={this.props.showArchitectures}
             closeHistoryPanel={this.props.closeHistoryPanel}
-            getReleaseHistory={this.props.getReleaseHistory}
           />
         </div>
       </div>
@@ -27,6 +27,7 @@ export default class HistoryPanel extends Component {
 
 HistoryPanel.propTypes = {
   // state
+  releases: PropTypes.array.isRequired,
   revisionsMap: PropTypes.object.isRequired,
   releasedChannels: PropTypes.object.isRequired,
   revisionsFilters: PropTypes.object,
@@ -34,6 +35,5 @@ HistoryPanel.propTypes = {
   showArchitectures: PropTypes.bool,
   // actions
   selectRevision: PropTypes.func.isRequired,
-  closeHistoryPanel: PropTypes.func.isRequired,
-  getReleaseHistory: PropTypes.func.isRequired
+  closeHistoryPanel: PropTypes.func.isRequired
 };
