+++ static/js/publisher/release/releasesTable.js
@@ -13,6 +13,8 @@ import ChannelMenu from "./channelMenu";
 import PromoteButton from "./promoteButton";
 import HistoryPanel from "./historyPanel";
 
+import { getTrackingChannel } from "./releasesState";
+
 function getChannelName(track, risk) {
   return risk === UNASSIGNED ? risk : `${track}/${risk}`;
 }
@@ -72,7 +74,12 @@ export default class ReleasesTable extends Component {
 
     const isChannelClosed = this.props.pendingCloses.includes(channel);
     const isPending = hasPendingRelease || isChannelClosed;
-    const trackingChannel = this.props.getTrackingChannel(track, risk, arch);
+    const trackingChannel = getTrackingChannel(
+      releasedChannels,
+      track,
+      risk,
+      arch
+    );
 
     const isUnassigned = risk === UNASSIGNED;
     const isActive =
@@ -325,6 +332,7 @@ export default class ReleasesTable extends Component {
     return (
       <HistoryPanel
         key="history-panel"
+        releases={this.props.releases}
         revisionsMap={this.props.revisionsMap}
         revisionsFilters={this.props.revisionsFilters}
         releasedChannels={this.props.releasedChannels}
@@ -332,7 +340,6 @@ export default class ReleasesTable extends Component {
         selectRevision={this.props.selectRevision}
         showArchitectures={!!showArchitectures}
         closeHistoryPanel={this.props.closeHistoryPanel}
-        getReleaseHistory={this.props.getReleaseHistory}
       />
     );
   }
@@ -512,6 +519,7 @@ export default class ReleasesTable extends Component {
 
 ReleasesTable.propTypes = {
   // state
+  releases: PropTypes.array.isRequired,
   revisionsMap: PropTypes.object.isRequired,
   archs: PropTypes.array.isRequired,
   tracks: PropTypes.array.isRequired,
@@ -533,9 +541,7 @@ ReleasesTable.propTypes = {
   undoRelease: PropTypes.func.isRequired,
   clearPendingReleases: PropTypes.func.isRequired,
   closeChannel: PropTypes.func.isRequired,
-  getTrackingChannel: PropTypes.func.isRequired,
   toggleHistoryPanel: PropTypes.func.isRequired,
   selectRevision: PropTypes.func.isRequired,
-  closeHistoryPanel: PropTypes.func.isRequired,
-  getReleaseHistory: PropTypes.func.isRequired
+  closeHistoryPanel: PropTypes.func.isRequired
 };
